Will not delete service or bundle until every user is unsubscribed

+++ Users/views.py
@@ -357,8 +357,12 @@ def add_bundle(request):
         bundle_form = bForm();
         bundle_form.bundles = Bundle.objects.all()
 
+        for bundles in bundle_form.bundles:
+            if bundles.deleted==False:
+                print(bundles)
 
-        return render(request, 'Users/add_bundles.html', {'bundle_form': bundle_form.bundles.all()})
+
+        return render(request, 'Users/add_bundles.html', {'bundle_form': bundle_form.bundles.filter(deleted=False)})
 
 
 @login_required
@@ -446,12 +450,16 @@ def add_services(request):
         service_form = DisplayForm()
 
         #passing to the HTML ALL the contents in the Service database
-        service_form.services = Service.objects.filter(deleted=False)
+        service_form.services = Service.objects.all()
+
         #bundle_form = bForm();
         #bundle_form.bundles = Bundle.objects.all()
+        for service in service_form.services:
+            if service.deleted==False:
+                print(service)
+        #print(service_form.services.filter(deleted=False))
 
-
-        return render(request, 'Users/add_services.html', {'service_form': service_form.services.all()})
+        return render(request, 'Users/add_services.html', {'service_form': service_form.services.filter(deleted=False)})
 
 @login_required
 def display_services(request):
@@ -675,8 +683,10 @@ def market_rep(request):
                 if services.name == package_name:
                     package = services
                     if((checkUsersforSubscribedService(services.name))):
-                        (Service.objects.filter(name=package_name)).deleted = True
-                        print((Service.objects.filter(name=package_name)).deleted)
+                        for service in (Service.objects.filter(name=package_name)):
+                            service.deleted=True
+                            print(service)
+                            service.save()
                     else:
                         print(package)
                         Service.objects.filter(name=package_name).delete()
@@ -702,10 +712,14 @@ def market_rep(request):
             print("deleting bundle: ", bundle_name)
             for bundle in Bundle.objects.all():
                 if bundle.name == bundle_name:
-                    package = bundle
-                    print(package)
-                    Bundle.objects.filter(name=bundle_name).delete()
-                    print("deleting bundle object with name: ", bundle_name)
+                    if((checkUserforSuscbibedBundle(bundle.name))):
+                        for bundles in (Bundle.objects.filter(name=bundle_name)):
+                            bundles.deleted=True
+                            print(bundles)
+                            bundles.save()
+                    else:
+                        Bundle.objects.filter(name=bundle_name).delete()
+                        print("deleting bundle object with name: ", bundle_name)
 
             current_user = UserProfile.objects.get(user=request.user)
             current_user.save()
@@ -937,12 +951,20 @@ def checkUsersforSubscribedService(ServicesName):
     for user in allUsers:
         for service in user.services.all():
             if(service.name == ServicesName):
+                #service.deleted=True
                 return True
 
     return False
 
-#def checkUserforSuscbibedBundle(BundleName):
+def checkUserforSuscbibedBundle(BundleName):
+    allUsers = UserProfile.objects.all().filter(is_Market=False, is_Service=False)
+    for user in allUsers:
+        for bundles in user.bundles.all():
+            if(bundles.name == BundleName):
+                #service.deleted=True
+                return True
 
+    return False
 #def validate_views(obj, ):
     #if render_loc!='':
         #return render_loc
